initial support for k8s stacks and resources
Implement a k8s Stack construct which synthesizes
a YAML into the cloud assembly that contains
all the resources defined within the current scope.

Implement "L1" resources for k8s Deployment and Service
and leverage the API spec for their props.

Write a README which describes how to use this
to create a simple hello world application.