chore: pin jsii to 1.1.0 due to bug in 1.2.0
In jsii 1.2.0 the way python code loads jsii tarballs has been changed and is no longer "portable" in a sense that we can harvest the python code into a user-defined location in the cdk8s project (e.g. `imports`).

Until this is fixed, we are pinning jsii to 1.1.0.

jsii bug: https://github.com/aws/jsii/issues/1501