feat(cli): support CRDs with apiVersion "apiextensions.k8s.io/v1" (#142)
Stop failing when a CRD uses the v1 api version. It is compatible with v1beta from our point of view and therefore can be imported.

Furthermore, ignore properties that are not camelCase since these can't be compiled to jsii. Users will have to use escape hatches for these.

Clean up some code and tests.