feat(plus): service.addDeployment()
Introduce `service.addDeployment()` which associates a deployment to a service. The method `deployment.expose()` now just delegates the work to this method.

Additionally, the logic that added the selector label to the deployment was moved from `deployment._toKube()` into `service.addDeployment()` and combined with the logic that installed the selector on the service (which is much more natural).

BREAKING CHANGE: `deployment.expose()` now takes `port` as a positional argument (before: `deployment.expose({ port })`, now: `deployment.expose(port)`).