fix(lib): fail to import octal numbers via include (and helm)
Due to a bug in the `yaml` library (https://github.com/eemeli/yaml/issues/205) which
prevents using octal numbers (e.g. `0775`) in included manifests, we are temporarily using `js-yaml` as a parser. We can't completely move to `js-yaml` for serializing because it doesn't support "always quote".

The state of YAML is horrific.

Fixes #348