feat(plus): expose service options in `expose()` (#357)
This PR expands the `Deployment expose()` method to allow users to fully specify options to the generated Service and Port.  Originally I created this so I could provide a user defined name for the Service through metadata.  In the process I also found the need to allow specifying the port details for name and protocol.