fix(plus-17): don't allow containers to be contructed from containers (#404)
Add a check for the case where a container is attempted to be built from another container.

Any place that takes ContainerProps actually also accepts Containers. This happens because ContainerProps has all optional properties except image .  Since Container has an a public property image, it gets passed along but that is really the only thing that gets used. 

Problem is that if you pass in a container object, the nested container will not get all the settings, so I don't think this should be allowed.

note: this is a bit of an ugly hack to prevent this.  I was hoping to find a way to do this with the type system but didn't find a solid option without changing some public property names and breaking the API.

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*