chore: projen subprojects
Convert packages to use projen subprojects. This means that now we have a single projen entry point in the repo (the root `.projenrc.js`) which references `.subprojenrc.js` files within each module.

This is the first step in moving the project to projen completely. The remaining work is to codify the various repo-level configuration (workflows, etc) as part of the root project.