fix(cli): importing local files is broken on windows (#427)
Previously protocol was used to detect if file is local or on a remote url, however windows absolute file paths start with C: or D: etc. so are incorrectly detected as urls leading to error on line 46. fs.lstatSync(url).isFile() seemed to be the best option from what I can tell, but fs.existsSync() or others may be better. Thanks in advance for the library and any feedback on this PR. .

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*