Update imports so they will work with TS esModuleInterop
The imports as previously defined would run into an error when
imported by projects that use the current tsconfig.json default
of true for compilerOptions esModuleInterop.

Type originates at this import. A namespace-style import cannot
be called or constructed, and will cause a failure at runtime.
Consider using a default import or import require here instead.

This changes the code to use a require and thus allow usage with
projects using interop.