fix: set yaml default schema to 1.1 (#505)
*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*

---
Currently, the output of multiline strings is very ugly in the rendered YAML files; there are many extra newlines and backslashes, majorly hurting readability.

To fix #172, this PR returns to the default YAML string settings that allow the parser to decide between double-quoting or creating a block-literal from a string. 

To ensure we don't bring back the issues from #325, this PR also sets the default YAML schema to `yaml-1.1`, to retain compatibility with most other YAML parsers' default settings.

Aside from the tests in the PR, I have also confirmed this fix against the repro steps in #325.