fix(website): TS getting started is not clear enough with relation to synth (#708)
Currently, the instructions for TS mention force the user to run a watch command in order to properly synth the app.

<img width="714" alt="Screen Shot 2021-10-06 at 9 12 36 PM" src="https://user-images.githubusercontent.com/1428812/136259906-ae278b29-03ce-469e-8edb-f2bc328bcf93.png">

This instruction is easily overlooked since its not intuitive, in addition, it creates a complexity of having to run a background process. 

Skipping this command will make it so running `cdk8s synth` further down the guide will have no affect. 

Instead, we instruct the user to run a `npm run compile && cdk8s synth` command in order to ensure compilation happens. As for running watch, we now mention this in a quote as an option to forgo explicit compilation - but this is not the main path.

Here is how TS will look like following this PR:

<img width="442" alt="Screen Shot 2021-10-06 at 9 06 49 PM" src="https://user-images.githubusercontent.com/1428812/136260219-e70b926c-fa75-47f8-b87e-7ea5802c84f6.png">
<img width="429" alt="Screen Shot 2021-10-06 at 9 07 18 PM" src="https://user-images.githubusercontent.com/1428812/136260235-badfac92-a195-4f3e-a185-de53234124ea.png">

In contrast with python for example, where we simply use `cdk8s synth`.

<img width="431" alt="Screen Shot 2021-10-06 at 9 09 02 PM" src="https://user-images.githubusercontent.com/1428812/136260312-7f1db9b0-935d-4a3a-a285-f299ce40dbc4.png">
<img width="434" alt="Screen Shot 2021-10-06 at 9 09 20 PM" src="https://user-images.githubusercontent.com/1428812/136260320-288afaf6-a220-4fd3-a3f3-568f6750b044.png">