fix(docs): lots of clean up and style enhancements (#963)
## Summary

This PR includes some small style and navigation changes that hopefully make the doc site easier to use.

Some of the changes
- Lifted contents out of /projects
- Renamed "Assets" to "Media" and made it a single page
- Reordered left nav tree
- Combined examples into a single page
- Redesigned /reference landing page
- Configured "Overview" pages to be hidden from nav and act as landing pages for the related section.
- Removed duplicate title on left nav
- Added shortcut key hint to search bar
- Renamed /concepts pages to /basics (this will break user's bookmarks, so i'd like some feedback on this)
- Added Stack overflow icon to footer
- Support for section in URL (as hash) so that back nav works as expected.
- Renamed title of right nav menu from "Table of contents" to "On this page"
- Changed header's repo icon from a git icon to github icon
- Renamed site title from "cdk8s Documentation" to "cdk8s"
- Linked all tab-able content boxes so that when users select a language in one, it updates all tab-able content boxes on the site.
- Replaced use of "notice" admonitions with "info" since "notice" is not a valid mkdocs [admonitions type](https://squidfunk.github.io/mkdocs-material/reference/admonitions/#supported-types).


Demo of changes
https://share.cleanshot.com/VMSEcZ

What I wanted to fix but couldn't

The API reference pages are going to be difficult to style. The sections are labelled with typical markdown header syntax (i.e. `# Construct`). It looks like every header is generated with an anchor tag though so I may be able to use a css selector to grab the header. This isn't ideal though and could lead to some pretty hard to debug style issues down the road.

Haven't found a clear way to modify the right nav item order on API reference pages. Because it's ordered by header hierarchy we may not have many options. I've yet to find an escape hatch for this in mkdocs.

## Deploy notes

⚠️ I recommend merging https://github.com/cdk8s-team/cdk8s/pull/964 first.