chore: k8s-25/main branch and all corresponding wiring (#1006)
## Summary

This PR adds K8s-25 to the repo. I didn't update the docs because we're moving the plus/ docs to the cdk8s-plus repo.

Issue:
- https://github.com/cdk8s-team/cdk8s-plus/issues/1131

Parent branch PR for moving docs/plus files from cdk8s repo to cdk8s-plus repo:

- https://github.com/cdk8s-team/cdk8s/pull/1005

## Merge order

A few of my PRs touch a lot of the same code so i've documented a merge strategy. See:

- https://github.com/cdk8s-team/cdk8s-plus/pull/1224

Signed-off-by: Ryan Parker <ryan.parker3@outlook.com>