chore(docs/plus): move cdk8s+ docs to cdk8s-plus repo (#1005)
## Summary

This PR moves the CDK8s+ docs to the CDK8s+ repo. Updates the `docs/build.sh` to copy the files to the versioned `docs/plus/$version/` sub dir once they are published.

Lets merge the CDK8s+ docs first so that we can test the `docs/build.sh` before merging: 
- https://github.com/cdk8s-team/cdk8s-plus/pull/1224

Issue: 
- https://github.com/cdk8s-team/cdk8s-plus/issues/493


Child PRs:
- https://github.com/cdk8s-team/cdk8s/pull/1006

## Merge order

A few of my PRs touch a lot of the same code so i've documented a merge strategy. See:

- https://github.com/cdk8s-team/cdk8s-plus/pull/1224